test(koa-graphql): use graphql.query for testing
