+++ server/koa-graphql/src/__tests__/index.js
@@ -2,12 +2,14 @@
 
 import path from 'path';
 
+import Koa from 'koa';
+import getPort from 'get-port';
+import fetch, { type Response as ResponseType } from 'node-fetch';
 import { outputFileSync } from 'output-file-sync';
 
 import Graphql from '../index';
 
-import runServer from './__ignore__/runServer';
-
+const graphqlFolder = path.resolve(__dirname, './__ignore__/schema');
 const additionalSchema = {
   typeDefs: `
   extend type Query {
@@ -33,11 +35,9 @@ const requestResult = {
   },
 };
 
-let runningServer: http$Server;
-
 describe('graphql', () => {
   test('relay', () => {
-    const graphql = new Graphql(path.resolve(__dirname, './__ignore__/schema'));
+    const graphql = new Graphql(graphqlFolder);
 
     outputFileSync.destPaths = [];
     outputFileSync.contents = [];
@@ -65,14 +65,24 @@ type User {
     ]);
   });
 
-  describe('middleware', () => {
-    test('basic usage', async () => {
-      const { server, request } = await runServer();
+  test('middleware', async () => {
+    const app = new Koa();
+    const port = await getPort();
+    const graphql = new Graphql(graphqlFolder);
 
-      runningServer = server;
+    app.use(graphql.middleware());
 
-      expect(
-        await request(`
+    const server = app.listen(port);
+
+    expect(
+      await fetch(`http://localhost:${port}`, {
+        method: 'POST',
+        headers: {
+          'Content-Type': 'application/json',
+          Accept: 'application/json',
+        },
+        body: JSON.stringify({
+          query: `
   {
     version
     users {
@@ -83,76 +93,61 @@ type User {
       }
     }
   }
-        `),
-      ).toEqual({
-        data: {
-          version: '1.0.0',
-          users: [
-            {
-              id: 'user-id',
-              event: {
-                id: 'event-id',
-                name: 'event-name',
-              },
+        `,
+        }),
+      }).then((res: ResponseType) => res.json()),
+    ).toEqual({
+      data: {
+        version: '1.0.0',
+        users: [
+          {
+            id: 'user-id',
+            event: {
+              id: 'event-id',
+              name: 'event-name',
             },
-          ],
+          },
+        ],
+      },
+    });
+
+    server.close();
+  });
+
+  test.each`
+    info        | option
+    ${'string'} | ${additionalSchema}
+    ${'array'}  | ${{ ...additionalSchema, typeDefs: [additionalSchema.typeDefs] }}
+  `(
+    'additional schema with $info',
+    async ({ option }: {| option: typeof additionalSchema |}) => {
+      expect(
+        await new Graphql(graphqlFolder, option).query(requestQuery),
+      ).toEqual(requestResult);
+    },
+  );
+
+  test.each`
+    filePath                                          | isEqual
+    ${'./__ignore__/schemaChanged/key.js'}            | ${true}
+    ${'./__ignore__/schemaChanged/emptyResolvers.js'} | ${false}
+    ${'./notIncludePath.js'}                          | ${false}
+  `(
+    'update resolver when file is changed with filePath = $filePath',
+    async ({ filePath, isEqual }: {| filePath: string, isEqual: boolean |}) => {
+      const graphql = new Graphql(graphqlFolder, additionalSchema);
+
+      expect(await graphql.query(requestQuery)).toEqual(requestResult);
+
+      graphql.update(path.resolve(__dirname, filePath));
+
+      const result = await graphql.query(requestQuery);
+
+      (isEqual ? expect(result) : expect(result).not).toEqual({
+        data: {
+          key: 'test',
         },
       });
-    });
-
-    test('additional schema with string', async () => {
-      const { server, request } = await runServer(additionalSchema);
-
-      runningServer = server;
-
-      expect(await request(requestQuery)).toEqual(requestResult);
-    });
-
-    test('additional schema with array', async () => {
-      const { server, request } = await runServer({
-        ...additionalSchema,
-        typeDefs: [additionalSchema.typeDefs],
-      });
-
-      runningServer = server;
-
-      expect(await request(requestQuery)).toEqual(requestResult);
-    });
-
-    test.each`
-      filePath                                          | isEqual
-      ${'./__ignore__/schemaChanged/key.js'}            | ${true}
-      ${'./__ignore__/schemaChanged/emptyResolvers.js'} | ${false}
-      ${'./notIncludePath.js'}                          | ${false}
-    `(
-      'update resolver when file is changed with filePath = $filePath',
-      async ({
-        filePath,
-        isEqual,
-      }: {|
-        filePath: string,
-        isEqual: boolean,
-      |}) => {
-        const { server, graphql, request } = await runServer(additionalSchema);
-
-        runningServer = server;
-
-        expect(await request(requestQuery)).toEqual(requestResult);
-
-        graphql.update(path.resolve(__dirname, filePath));
-
-        const result = await request(requestQuery);
-
-        (isEqual ? expect(result) : expect(result).not).toEqual({
-          data: {
-            key: 'test',
-          },
-        });
-      },
-    );
-
-    afterEach(() => {
-      runningServer.close();
-    });
-  });
+    },
+  );
 });
